Make the code a bit more generic (trivial). Different Super I/Os use different config ports.
Acked-by: Uwe Hermann <uwe@hermann-uwe.de>



git-svn-id: svn://svn.coreboot.org/coreboot/trunk@2782 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1

+++ util/superiotool/smsc.c
@@ -53,7 +53,7 @@ void dump_smsc(uint32_t port, uint32_t id)
 	}
 }
 
-void probe_idregs_smsc(uint32_t port)
+void probe_idregs_smsc(unsigned short port)
 {
 	uint16_t id, rev;
